Fix stripe errors and controller
