+++ src/main/java/com/castletroymedical/stripe/controller/PaymentController.java
@@ -7,6 +7,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.castletroymedical.stripe.dto.CreatePayment;
 import com.castletroymedical.stripe.dto.CreatePaymentResponse;
+import com.stripe.exception.StripeException;
+import com.stripe.model.PaymentIntent;
+import com.stripe.param.PaymentIntentCreateParams;
 
 
 @RestController
